dbus: ensure expected bytes width of DBus numbers
t - UINT64 - guint64 (unsigned long is not wide enough on 32bit architectures)

Resolves: rhbz#1256456

+++ src/dbus/abrt-dbus.c
@@ -625,8 +625,8 @@ static void handle_method_call(GDBusConnection *connection,
 
             g_variant_builder_add(response_builder, "{s(its)}",
                                                     element_name,
-                                                    element_info->flags,
-                                                    size,
+                                                    (gint32)element_info->flags,
+                                                    (guint64)size,
                                                     element_info->content);
         }
